Refactor and implement idea for dropdown sub-headers.
